+++ src/routes/(app)/+layout.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
   import "../../app.css";
-  import Drawer from "./Drawer.svelte";
-  import Footer from "./Footer.svelte";
-  import Navbar from "./Navbar.svelte";
-  import Toast from "./Toast.svelte";
+  import Drawer from "../Drawer.svelte";
+  import Footer from "../Footer.svelte";
+  import Navbar from "../Navbar.svelte";
+  import Toast from "../Toast.svelte";
 </script>
 
 <nav class="contents">
