+++ src/routes/DsekLogo.svelte
