+++ src/routes/LanguageSwitcher.svelte
