+++ src/routes/Navbar.svelte
