+++ src/routes/NotificationBell.svelte
