+++ src/routes/Toast.svelte
